+++ crates/swc_ecma_lints/tests/fixture.rs
@@ -69,7 +69,8 @@ fn pass(input: PathBuf) {
             Program::Script(s) => {
                 rules.lint_script(s);
             }
-            _ => unreachable!(),
+            // TODO: reenable once experimental_metadata breaking change is merged
+            //_ => unreachable!(),
         });
 
         if handler.has_errors() {